Add a controller to interact with a live database
Because Mongoose is a tested library with such useful wrappers, unit testing route handlers
is tedious and mostly redundant.  Instead, we will integration-test routes to alert us to
any changes to api output that might break front-end code.  The testDBController class provides
a number of convenient utilities for adding mock data to a live mongo database and interacting with/
asserting on that data.